pyproject.toml
[project]
name = "tap-zoomphone"
version = "0.0.1"
description = "Singer tap for ZoomPhone, built with the Meltano Singer SDK."
readme = "README.md"
authors = [{ name = "Chad Campling", email = "chad.campling@opteonsolutions.com" }]
keywords = [
    "ELT",
    "ZoomPhone",
]
classifiers = [
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
license = "Apache-2.0"
license-files = [ "LICENSE" ]
requires-python = ">=3.10"
dependencies = [
    "singer-sdk[faker]~=0.50.1",
    "requests~=2.32.3",
    "python-dateutil>=2.9.0.post0",
]

[project.optional-dependencies]
s3 = [
    "fs-s3fs~=1.1.1",
]

[project.scripts]
# CLI declaration
tap-zoomphone = 'tap_zoomphone.tap:TapZoomPhone.cli'

[dependency-groups]
dev = [
    { include-group = "test" },
]
test = [
    "pytest>=8",
    "pytest-github-actions-annotate-failures>=0.3",
    "singer-sdk[testing]",
]

[tool.pytest.ini_options]
addopts = [
    "--durations=10",
]

[tool.mypy]
python_version = "3.12"
warn_unused_configs = true

[tool.ruff]
target-version = "py39"

[tool.ruff.lint]
ignore = [
    "COM812",  # missing-trailing-comma
]
select = ["ALL"]

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[build-system]
requires = [
    "hatchling>=1,<2",
]
build-backend = "hatchling.build"
